requirements.txt
Django==1.5.4
Fabric==1.8.1
PIL==1.1.7
South==0.8.2
Unipath==1.0
django-appconf==0.6
django-braces==1.2.2
django-classy-tags==0.4
django-social-auth==0.7.28
django-sslify==0.2
django-twitter-tag==1.2
ecdsa==0.10
feedparser==5.1.3
httplib2==0.8
oauth2==1.5.211
paramiko==1.12.0
psycopg2==2.5.2
pycparser==2.10
pycrypto==2.6.1
python-openid==2.2.5
requests==2.1.0
selenium==2.38.1
six==1.2.0
twitter==1.9.1
wsgiref==0.1.2